feat: add RegisterClubForm and use temporary storage to register club
