+++ packages/web/src/common/components/Modal/RestoreDraftModal.tsx
@@ -6,12 +6,14 @@ import CancellableModalContent from "./CancellableModalContent";
 
 interface RestoreDraftModalProps {
   isOpen: boolean;
+  mainText?: string;
   onConfirm: () => void;
   onClose: () => void;
 }
 
 const RestoreDraftModal: React.FC<RestoreDraftModalProps> = ({
   isOpen,
+  mainText = undefined,
   onConfirm,
   onClose,
 }) => (
@@ -22,7 +24,7 @@ const RestoreDraftModal: React.FC<RestoreDraftModalProps> = ({
       onConfirm={onConfirm}
       onClose={onClose}
     >
-      임시 저장된 데이터가 있습니다. 불러오시겠습니까?
+      {mainText ?? `임시 저장된 데이터가 있습니다. 불러오시겠습니까?`}
       <Typography color="GRAY.600" fs={12} lh={16} fw="REGULAR">
         ‘새로 작성하기’를 선택하면 기존 임시 저장된 데이터가 삭제됩니다.
       </Typography>
