+++ packages/web/src/constants/localStorage.ts
@@ -1,3 +1,4 @@
 export const LOCAL_STORAGE_KEY = {
   ACTIVITY_REPORT: "activity-report",
+  REGISTER_CLUB: "register-club",
 } as const;
